requirements.txt
requests>=2.31.0
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
cohere>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
